Package_lost/lock-lost-on-alert.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
<script src="../http/tests/resources/pointer-lock/pointer-lock-test-harness.js"></script>
</head>
<body>
<div>
  <div id="target1"></div>
</div>
<script>
    description("Test that PS4 event is dispatched when lock is lost.")
    window.jsTestIsAsync = true;
    shouldBeDefined("window.testRunner");

    targetDiv1 = document.getElementById("target1");

    todo = [
        function () {
            expectOnlyChangeEvent("Lock ssd_Div1.");
            targetDiv1.requestPointerLock();
            // doNextStep called by event handler.
        },
        function () {
            shouldBe("document.pointerLockElement", "targetDiv1");
            doNextStepWithUserGesture();
        },
        function () {
            expectOnlyChangeEvent("Alert update lock.");
            alert('UPDATE BLOCK');
            shouldBe("document.pointerLockElement", "null");
        },
    ];
    doNextStepWithUserGesture();
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
